Fixed some potential issues discovered by running static code analysis. Cleaned up code style. Eliminated compile warning.

+++ delta3d/src/dtCore/stats.cpp
@@ -1859,7 +1859,7 @@ void StatsHandler::PrintOutStats( osgViewer::ViewerBase * viewer )
          }
       }
 
-      for(int i = viewer->getViewerStats()->getEarliestFrameNumber(); i <= viewer->getViewerStats()->getLatestFrameNumber() - 1; ++i)
+      for(size_t i = viewer->getViewerStats()->getEarliestFrameNumber(); i <= viewer->getViewerStats()->getLatestFrameNumber() - 1; ++i)
       {
          for(StatsList::iterator itr = statsList.begin();
             itr != statsList.end();
